refactor(agentplugin): move ListTables filter registry to libs/go-common
The registry was originally placed under
services/agent/internal/discovery, which makes the ListTables filter
hook exclusively reachable from the agent service's own packages —
plugins outside services/agent (the entire decisionbox-enterprise
repo, future custom builds) couldn't use it.

Moving the registry to libs/go-common/agentplugin alongside the
context-provider and ask-override registries puts every plugin hook
in one public, importable place. The pre-move discovery-package
import path is preserved via thin aliases (RegisterListTablesFilter,
ApplyListTablesFilters, ResetListTablesFiltersForTest) so existing
in-package callers keep compiling and tests stay in place.
